完成权限管理功能和界面以及部分细节优化

+++ Model-Service/src/main/java/com/cn/lucky/morning/model/service/impl/AuthorityServiceImpl.java
@@ -70,11 +70,7 @@ public class AuthorityServiceImpl implements AuthorityService {
         }
         List<Authority> list = Lists.newArrayList();
         for (Long id : ids) {
-            Authority auth = (Authority) cacheService.get(Const.cache.AUTHORITY_ID + id);
-            if (auth == null) {
-                auth = authorityMapper.selectByPrimaryKey(id);
-                cacheService.set(Const.cache.AUTHORITY_ID + id, auth, Const.cache.AUTHORITY_ID_TTL);
-            }
+            Authority auth =getById(id);
             list.add(auth);
         }
         return list;
@@ -87,7 +83,7 @@ public class AuthorityServiceImpl implements AuthorityService {
         }
         for (Long id : ids) {
             if (id != null) {
-                if (authorityMapper.deleteByPrimaryKey(id) == 0) {
+                if (!delete(id)) {
                     return false;
                 }
             }
@@ -97,11 +93,17 @@ public class AuthorityServiceImpl implements AuthorityService {
 
     @Override
     public boolean delete(Long id) {
+        Authority authority = getById(id);
+        if (authority == null){
+            return true;
+        }
+        cacheService.del(Const.cache.AUTHORITY_ID+id);
         return authorityMapper.deleteByPrimaryKey(id) > 0;
     }
 
     @Override
     public boolean edit(Authority authority) {
+        cacheService.del(Const.cache.AUTHORITY_ID+authority.getId());
         return authorityMapper.updateByPrimaryKeySelective(authority) > 0;
     }
 
@@ -129,6 +131,11 @@ public class AuthorityServiceImpl implements AuthorityService {
 
     @Override
     public Authority getById(Long id) {
+        Authority authority = (Authority) cacheService.get(Const.cache.AUTHORITY_ID+id);
+        if (authority == null){
+            authority = authorityMapper.selectByPrimaryKey(id);
+            cacheService.set(Const.cache.AUTHORITY_ID+id,authority,Const.cache.AUTHORITY_ID_TTL);
+        }
         return authorityMapper.selectByPrimaryKey(id);
     }
 
@@ -146,7 +153,15 @@ public class AuthorityServiceImpl implements AuthorityService {
     public boolean deleteByAuthorityGroupId(Long authorityGroupId) {
         AuthorityExample example = new AuthorityExample();
         example.createCriteria().andGroupIdEqualTo(authorityGroupId);
-        return authorityMapper.deleteByExample(example)>0;
+        List<Authority> list = authorityMapper.selectByExample(example);
+        if (list!=null){
+            for (Authority authority : list){
+                if (!delete(authority.getId())){
+                    return false;
+                }
+            }
+        }
+        return true;
     }
 
 }
